+++ dflib/src/main/java/com/nhl/dflib/concat/HConcatDataFrame.java
@@ -39,7 +39,7 @@ public class HConcatDataFrame implements DataFrame {
     }
 
     @Override
-    public long height() {
+    public int height() {
         return leftSource.height();
     }
